requirements.txt
pandas
requests
beautifulsoup4
selenium
mysql-connector-python
python-dotenv
